--- mupl/utils/__init__.py
@@ -1,1 +1,0 @@
-from mupl.utils.logs import setup_logs
