+++ src/loc/fr.json
