+++ src/loc/vi.json
@@ -1,0 +1,54 @@
+{
+    "invalid_folder_to_upload": "Không tìm thấy tệp hợp lệ để tải lên, đang thoát.",
+    "zips_no_manga_id_skip_message": "Bỏ qua các tệp {} vì thiếu id manga",
+    "zips_no_manga_id_skip_message_logs": "{}, hãy kiểm tra nhật ký để tìm tên tệp.",
+    "check_file_name_to_id": "Vui lòng kiểm tra tệp tên-id của bạn.",
+    "uploading_draft": "Đang tải lên",
+    "finish_upload": "Hoàn thành tải lên",
+    "keyboard_interrupt_exit": "Phát hiện ngắt bàn phím, đang thoát.",
+    "failed_uploads": "Tải lên thất bại:",
+    "upload_method_folder": "Thư mục",
+    "upload_method_archive": "Lưu trữ",
+
+    "new_update_found": "Tìm thấy bản cập nhật: {} => {}: {}.",
+    "new_update_warning": "Phiên bản mới có thể có những thay đổi không tương thích, vui lòng kiểm tra trang phát hành trên GitHub để xem danh sách các thay đổi.",
+    "not_updating": "Không cập nhật.",
+    "input_want_update": "Bạn có muốn cập nhật không? [y/N] ",
+    "successfully_updated": "Đã cập nhật, khởi động lại trình tải lên để chạy phiên bản mới.",
+    "update_error": "Không thể cập nhật.",
+
+    "keyboard_interrupt_cancel": "Phát hiện ngắt bàn phím. Hủy các tác vụ...",
+    "chapter_number_manga": "Chương",
+    "volume_number_manga": "Volume",
+    "chapter_title_manga": "Tiêu đề",
+    "language_manga": "Ngôn ngữ",
+    "groups_manga": "Nhóm",
+    "publish_date_manga": "Đăng vào",
+    "invalid_images_to_upload": "Không tìm thấy hình ảnh hợp lệ để tải lên, đang bỏ qua.",
+    "draft_create_session": "Đã tạo phiên tải lên: {}",
+    "images_to_upload": "{} hình ảnh cần tải lên.",
+    "threaded_upload_running": "Đang chạy trình tải lên đa luồng.",
+    "non_threaded_upload_running": "Đang chạy trình tải lên đơn luồng.",
+    "draft_deleting_failed_upload": "Xóa bản nháp do tải hình ảnh lên thất bại.",
+    "not_defined_value": "Không có giá trị",
+
+    "uploading_images": "Đang tải hình ảnh {} đến {}.",
+    "uploading_images_error": "Lỗi tải hình ảnh {} đến {}, thử {}/{}.",
+    "successful_upload_message": "Thành công: Đã tải lên trang {}, kích thước: {} mb.",
+    "error_create_draft_session": "Không thể tạo phiên tải lên cho {}.",
+    "uploading_successfully": "Đã tải lên thành công: {}, {}.",
+    "uploading_failed": "Không thể tải lên {}",
+    "image_split": "{} có kích thước trên 10000px {}, đã chia thành {} ảnh.",
+    "image_split_not_defined": "Hình ảnh {} vượt quá 10,000px ở {}. Bạn đã thêm ID manga vào bản đồ ID để xác định nó là {} và cần được chia tách chưa?",
+    "image_combine": "Đang kết hợp `{}` với `{}` vì `{}` nhỏ hơn {}px.",
+    "image_skip": "Bỏ qua `{}` vì nó nhỏ hơn {}px và không khớp {} của hình ảnh trước.",
+
+
+    "naming_format_incorrect": "{} không đúng định dạng tên, bỏ qua.",
+    "skip_no_manga_id": "Đã bỏ qua {}, không tìm thấy id manga.",
+
+    "not_logged_in": "401: Chưa đăng nhập.",
+    "logged_in": "Đã đăng nhập.",
+
+    "unable_convert_api_response_to_json": "{}: Không thể chuyển đổi phản hồi API của mangadex thành JSON."
+}
